% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postgres.R
\name{postgres.connect.pgpass}
\alias{postgres.connect.pgpass}
\title{PostgreSQL Connection Function Using .pgpass File}
\usage{
postgres.connect.pgpass()
}
\value{
A connection object (\code{con}) to the PostgreSQL database.
}
\description{
This function establishes a connection to a PostgreSQL database using the parameters
specified in the .pgpass file. It allows you to connect without providing a password
separately.
}
\examples{
# Connect to a PostgreSQL database using .pgpass file
con <- postgres.connect.pgpass()
devtools::document()
}
