% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postgres.R
\name{postgres.connect}
\alias{postgres.connect}
\title{PostgreSQL Connection Function}
\usage{
postgres.connect(
  .database = NULL,
  .port = 5432,
  .host = "localhost",
  .user = "postgres",
  .password = "password"
)
}
\arguments{
\item{.database}{Name of the database to connect to (default is NULL). This parameter will be converted to lowercase.}

\item{.port}{Port number of the PostgreSQL server (default is 5432).}

\item{.host}{Hostname of the PostgreSQL server (default is 'localhost').}

\item{.user}{Username to authenticate with (default is 'postgres').}

\item{.password}{Password for the provided username (default is 'password').}
}
\value{
A connection object (\code{con}) to the specified PostgreSQL database.
}
\description{
This function establishes a connection to a PostgreSQL database using the provided
parameters. It allows you to set the database name, port, host, user, and password.
}
\examples{
# Connect to a local PostgreSQL database
con <- postgres.connect(.database = "mydatabase", .user = "myuser", .password = "mypassword")

}
