% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postgres.R
\name{postgres.delete}
\alias{postgres.delete}
\title{Delete Rows from PostgreSQL Table Based on Conditions}
\usage{
postgres.delete(.conn, .tablename, .predicatelist)
}
\arguments{
\item{.conn}{A connection object to the PostgreSQL database.}

\item{.tablename}{Name of the table in the PostgreSQL database from which rows will be deleted.}

\item{.predicatelist}{A named list where names represent the column names and the list elements are the values used in the WHERE clause for deletion.}
}
\value{
A string representation of the WHERE clause used for the deletion.
}
\description{
This function deletes rows from a specified PostgreSQL table based on conditions provided in a predicate list.
}
\examples{
# Assuming you have an active connection:
# clause <- postgres.delete(.conn, .tablename = "mytable", .predicatelist = list(column1 = "value1", column2 = "value2"))
# print(clause)

}
