% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postgres.R
\name{postgres.log.apiRequest}
\alias{postgres.log.apiRequest}
\title{Log API Request in PostgreSQL}
\usage{
postgres.log.apiRequest(
  .name,
  .url,
  .body,
  .response,
  .tablename = "api_request_logs"
)
}
\arguments{
\item{.name}{Name of the API request}

\item{.url}{URL of the API request}

\item{.body}{Body of the API request}

\item{.response}{Response from the API}

\item{.tablename}{Name of the table to log the data, default is "api_request_logs"}
}
\description{
This function logs API request and response data to a PostgreSQL database.
}
