% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postgres.R
\name{postgres.update}
\alias{postgres.update}
\title{Update Rows in a PostgreSQL Table Based on Conditions}
\usage{
postgres.update(.conn, .tablename, .predicatelist, .updatelist)
}
\arguments{
\item{.conn}{A connection object to the PostgreSQL database.}

\item{.tablename}{Name of the table in the PostgreSQL database where rows will be updated.}

\item{.predicatelist}{A named list where names represent the column names and the list elements are the values used in the WHERE clause for updating.}

\item{.updatelist}{A named list where names represent the column names and the list elements are the new values to set.}
}
\value{
A string representation of the SQL UPDATE statement executed.
}
\description{
This function updates rows in a specified PostgreSQL table based on conditions provided in a predicate list
and changes specified in an update list.
}
\examples{
# Assuming you have an active connection:
# statement <- postgres.update(.conn, .tablename = "mytable", .predicatelist = list(id = 1), .updatelist = list(name = "new_name"))
# print(statement)

}
